Strategix & Cetautomatix reformating finished! 🎉🎉🎉 Available actions: both Phares & Pavillon
