+++ src/assurancetourix/strategix/strategix/action_objects/__init__.py
