+++ src/assurancetourix/strategix/strategix/action_objects/gobelet.py
